ДЗ2. Рефакторинг
