+++ scripts/catalog_functions/getElements.js
@@ -1,0 +1,16 @@
+export const getElements = (parentElement) => {
+    const elements = {};
+
+    const renderParams = {
+        nameElement: ".item-name",
+        priceElement: ".item-price",
+        photoElement: ".item-photo",
+        favButtonElement: ".favorite-btn",
+    };
+
+    Object.entries(renderParams).map(([key, value]) => {
+        elements[key] = parentElement.querySelector(value);
+    });
+
+    return elements;
+};
